enhancement: ttl handlers moved to service layer
