+++ src/test.rs
@@ -4,13 +4,13 @@ use bytes::BytesMut;
 
 use crate::{
     adapters::in_memory::InMemoryDb,
-    backgrounds::{delete_actor, set_ttl_actor},
     config::Config,
     services::{
         config_handler::ConfigHandler,
         interface::{Database, TRead, TWriteBuf},
         parser::MessageParser,
         persistence_handler::PersistenceHandler,
+        ttl_handlers::{delete::delete_actor, set::set_ttl_actor},
         ServiceFacade,
     },
 };
